Update nginx and install script

+++ install.sh
@@ -1,16 +1,31 @@
 #!/bin/bash
-set -e
+set -ex
+EXECUTION_DIR=$(pwd)
 
 pushd temp-chat-server
 npm install
 popd
 
-ln -s /temp-chat-deploy/temp-chat-nginx-config.conf /etc/nginx/sites-available/temp-chat-nginx-config.conf
+if [ -L /etc/nginx/sites-available/temp-chat-nginx-config.conf ]; then
+	rm /etc/nginx/sites-available/temp-chat-nginx-config.conf
+fi
 
-rm /etc/nginx/sites-enabled/default
+if [ -L /etc/nginx/sites-enabled/temp-chat-nginx-config.conf ]; then
+	rm /etc/nginx/sites-enabled/temp-chat-nginx-config.conf
+fi
 
-ln -s /etc/nginx/sites-available/temp-chat-nginx-config.conf /etc/nginx/sites-available/.
+ln -s $EXECUTION_DIR/temp-chat-nginx-config.conf /etc/nginx/sites-available/temp-chat-nginx-config.conf
 
-rm /var/www/html/*
+if [ -f /etc/nginx/sites-enabled ]; then
+	rm /etc/nginx/sites-enabled/default
+fi
 
-cp ~/temp-chat-deploy/temp-chat-client/dist/temp-chat-client/* /var/www/html/.
+ln -s /etc/nginx/sites-available/temp-chat-nginx-config.conf /etc/nginx/sites-enabled/.
+
+if ! [ -z "$(ls -A /var/www/html/)" ]; then
+	rm /var/www/html/*
+fi
+
+cp temp-chat-client/dist/temp-chat-client/* /var/www/html/.
+
+systemctl restart nginx
